+++ NEWS.md
@@ -1,6 +1,6 @@
 # anteo 0.0.3
 
-* Simple update with a new function: `lasso_extract` that extracts the list of variables selected in a LASSO regression from `cv.glmnet()` in the `glmnet` package.
+* Simple update with a new function: `cv_grplasso` which performs 10-fold cross-validation using `grplasso::grplasso`. Produces a data frame and a plot of the RMSE values depending on the tuning parameter lambda.
 
 # anteo 0.0.1